requirements-cloud-function.txt
attrs==19.3.0
backoff==1.10.0
blis==0.7.5
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.8
click==7.1.2
cryptography==3.4.8
cymem==2.0.6
future==0.18.2
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==1.42.1
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.2
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==4.7.4
murmurhash==1.0.6
numpy==1.21.4
packaging==21.3
pathy==0.6.1
plac==0.9.6
preshed==3.0.6
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pydantic==1.7.4
PyJWT==1.7.1
pyparsing==3.0.6
requests==2.26.0
rsa==4.8
six==1.14.0
smart-open==5.2.1
spacy==3.0.5
spacy-legacy==3.0.8
srsly==2.4.2
thinc==8.0.13
tqdm==4.62.3
typer==0.3.2
typing==3.7.4.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
yarl==1.4.2
zipp==3.6.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.0.0/en_core_web_md-3.0.0-py3-none-any.whl
